src/autocli/autocli/core/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
-e git+https://github.com/RakeemRanger/azurePythonautomation.git@929c61a37eacbee72c5bb7991fba64958ab08d29#egg=autocli
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-compute==34.1.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-network==29.0.0
azure-mgmt-resource==23.4.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
fastapi==0.115.12
Flask==3.1.1
guid==0.2.1
h11==0.16.0
httptools==0.6.4
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
logger==1.4
MarkupSafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
prettytable==3.16.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
